requirements.txt
pytest
Selenium
webdriver-manager
pytest-html
pytest-xdist
openpyxl
allure-pytest
